requirements.txt
django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
django-environ==0.12.0
drf-spectacular==0.28.0